% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Bryant2016}
\alias{Bryant2016}
\title{Bryant et al. (2016)}
\format{
A data frame with 299 rows and 8 variables
}
\source{
Bryant, B. R., Bryant, D. P., Porterfield, J., Dennis, M. S., Falcomata, T.,
Valentine, C., Brewer, C., & Bell, K. (2016).
The effects of a Tier 3 intervention on the mathematics performance of second grade
students with severe mathematics difficulties.
\emph{Journal of Learning Disabilities, 49}(2), 176-188.
doi:\href{https://doi.org/10.1177/0022219414538516}{10.1177/0022219414538516}
}
\description{
Data from a multi-level multiple baseline design conducted by
Bryant et al. (2016). The variables are as follows:
\itemize{
  \item \code{Study_ID}. Study identifier.
  \item \code{school}. School identifier.
  \item \code{case}. Student identifier.
  \item \code{session}. Measurement occasion.
  \item \code{treatment}. Indicator for treatment phase.
  \item \code{outcome}. Texas Early Mathematics Inventory (TEMI-Aim Check) scores.
  \item \code{trt_time}. Measurement occasion times treatment phase.
  \item \code{session_c}. Measurement occasion centered at the follow-up time (Measurement occasion 9).
}
}
\keyword{datasets}
