% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mlm.R
\name{g_mlm}
\alias{g_mlm}
\title{Calculates adjusted REML effect size}
\usage{
g_mlm(mod, p_const, r_const, infotype = "expected", returnModel = TRUE)
}
\arguments{
\item{mod}{Fitted model of class lmeStruct (estimated using \code{nlme::lme()})}

\item{p_const}{Vector of constants for calculating numerator of effect size.
Must be the same length as fixed effects in \code{mod}.}

\item{r_const}{Vector of constants for calculating denominator of effect size.
Must be the same length as the number of variance component parameters in \code{mod}.}

\item{infotype}{Type of information matrix. One of \code{"expected"} (the default),
\code{"observed"}, or \code{"average"}.}

\item{returnModel}{(Optional) If true, the fitted input model is included in the return.}
}
\value{
A list with the following components
\tabular{ll}{
\code{p_beta} \tab Numerator of effect size \cr
\code{r_theta} \tab Squared denominator of effect size \cr
\code{delta_AB} \tab Unadjusted (REML) effect size estimate \cr
\code{nu} \tab Estimated denominator degrees of freedom \cr
\code{J_nu} \tab Biased correction factor for effect size estimate \cr
\code{kappa} \tab Scaled standard error of numerator \cr
\code{g_AB} \tab Corrected effect size estimate \cr
\code{SE_g_AB} \tab Approximate standard error estimate \cr
\code{cnvg_warn} \tab Indicator that model did not converge \cr
\code{theta} \tab Estimated variance component parameters \cr
\code{info_inv} \tab Inversed information matrix \cr
}
}
\description{
Estimates a design-comparable standardized mean difference effect size based on data
from a multiple baseline design, using adjusted REML method as described in Pustejovsky, Hedges,
& Shadish (2014). Note that the data must contain one row per measurement occasion per case.
}
\examples{

library(nlme)
data(Laski, package = "scdhlm")
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
summary(Laski_RML)
g_mlm(Laski_RML, p_const = c(0,1), r_const = c(1,0,1), returnModel = FALSE)

data(Schutte, package = "scdhlm")
Schutte$trt.week <- with(Schutte, unlist(tapply((treatment=="treatment") * week,
         list(treatment,case), function(x) x - min(x))) + (treatment=="treatment"))
Schutte$week <- Schutte$week - 9
Schutte_RML <- lme(fixed = fatigue ~ week + treatment + trt.week,
                   random = ~ week | case,
                   correlation = corAR1(0.2, ~ week | case),
                   data = subset(Schutte, case != 4))
summary(Schutte_RML)
Schutte_g <- g_mlm(Schutte_RML, p_const = c(0,0,1,7), r_const = c(1,0,0,0,1))
summary(Schutte_g)
}
\references{
Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014).
Design-comparable effect sizes in multiple baseline designs: A general modeling framework.
\emph{Journal of Educational and Behavioral Statistics, 39}(4), 211-227. doi:\href{http://doi.org/10.3102/1076998614547577}{10.3102/1076998614547577}
}
